Commit before adding changes from other computer

+++ cryoS2Sdrop/cryoS2Sdrop/dataloader.py
@@ -436,7 +436,7 @@ class singleCET_FourierDataset(singleCET_dataset):
         if self.transform:
             subtomo, target, gt_subtomo = self.transform(subtomo, target, gt_subtomo)
             
-        return subtomo, target, gt_subtomo
+        return subtomo.float(), target.float(), gt_subtomo.float()
 
 
 class singleCET_ProjectedDataset(Dataset):
